Assignment 1 - Removing Projective Distortion/requirements.txt
tk
pillow
numpy
